"fix wrong size output, wrong targ tensor.type"

+++ engine.py
@@ -24,7 +24,7 @@ def train_fn(data_loader, model, optimizer, device, scheduler, epoch):
         ids = ids.to(device, dtype=torch.long)
         token_type_ids = token_type_ids.to(device, dtype=torch.long)
         mask = mask.to(device, dtype=torch.long)
-        targets = targets.to(device, dtype=torch.float)
+        targets = targets.to(device, dtype=torch.long)
 
         optimizer.zero_grad()
         outputs = model(ids=ids, mask=mask, token_type_ids=token_type_ids)
@@ -62,7 +62,7 @@ def eval_fn(data_loader, model, device):
             ids = ids.to(device, dtype=torch.long)
             token_type_ids = token_type_ids.to(device, dtype=torch.long)
             mask = mask.to(device, dtype=torch.long)
-            targets = targets.to(device, dtype=torch.float)
+            targets = targets.to(device, dtype=torch.long)
 
             outputs = model(ids=ids, mask=mask, token_type_ids=token_type_ids)
             loss = loss_fn(outputs, targets)
